+++ src/MSALWrapper/AuthFlow/Web.cs
@@ -66,9 +66,12 @@ namespace Microsoft.Authentication.MSALWrapper.AuthFlow
         /// <returns>A <see cref="Task"/> of <see cref="TokenResult"/>.</returns>
         public async Task<AuthFlowResult> GetTokenAsync()
         {
-            IAccount account = await this.pcaWrapper.TryToGetCachedAccountAsync(this.preferredDomain)
-                ?? null;
-            this.logger.LogDebug($"GetTokenNormalFlowAsync: Using account '{account.Username}'");
+            IAccount account = await this.pcaWrapper.TryToGetCachedAccountAsync(this.preferredDomain) ?? null;
+
+            if (account != null)
+            {
+                this.logger.LogDebug($"Using cached account '{account.Username}'");
+            }
 
             try
             {
@@ -142,24 +145,9 @@ namespace Microsoft.Authentication.MSALWrapper.AuthFlow
             return new AuthFlowResult(null, this.errors);
         }
 
-        private static HttpClient CreateHttpClient()
-        {
-            HttpClientHandler handler = new HttpClientHandler();
-
-            var client = new HttpClient(handler);
-
-            // Add default headers
-            client.DefaultRequestHeaders.CacheControl = new CacheControlHeaderValue
-            {
-                NoCache = true,
-            };
-
-            return client;
-        }
-
         private IPCAWrapper BuildPCAWrapper(ILogger logger, Guid clientId, Guid tenantId, string osxKeyChainSuffix)
         {
-            var httpFactoryAdaptor = new MsalHttpClientFactoryAdaptor(CreateHttpClient());
+            var httpFactoryAdaptor = new MsalHttpClientFactoryAdaptor();
             var clientBuilder =
                 PublicClientApplicationBuilder
                 .Create($"{clientId}")