+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed a bug where device code flow authentication would not use the file cache to first attempt to get a cached token silently, causing it to always prompt.
+
+### Removed
+- Removed sample projects that used the old `TokenFetcherPublicClient` api from the MSALWrapper project.
 
 ## [v0.2.0] - 2022-04-21
 ### Security